feat: Add endpoint for track of the week and update music.json file
This commit adds a new endpoint "/track-of-the-week" to retrieve the track of the week from the "music.json" file. It also updates the "music.json" file to include the track of the week data. This change allows users to fetch and view the track of the week on the application.

Closes #123